adding a restricted page for every user The host has the right to delete and update their room

+++ base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render,redirect
 from django.db.models import Q
 from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from django.contrib.auth import authenticate, login, logout
 from .forms import RoomForm
@@ -57,6 +59,7 @@ def room(request, pk):
     context = {'room': rooms}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -67,9 +70,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit this room.')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -80,9 +87,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
         room = Room.objects.get(id=pk)
+
+        if request.user != room.host:
+            return HttpResponse('You are not allowed to delete this room.')
+
         if request.method == 'POST':
             room.delete()
             return redirect('home')
